[proxy] use POST for all requests, send method as param
the current proxy behaves as a proxy should and sends the request to it's upstream counterpart in almost the exact way that it received it. These changes update the proxy to only run over POST and operate more like a API endpoint than a proxy. The endpoint now takes a URI, a http method, and a request body. It will combine those into a request to the URI and return the response just like the proxy did. This means that methods like "HEAD" don't need special treatment and has security benefits.